tredence_projects/backend/pyproject.toml
[tool.poetry]
name = "realtime-pair-backend"
version = "0.1.0"
description = "FastAPI backend for realtime pair programming prototype"
authors = ["Generated by ChatGPT"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
"uvicorn[standard]" = "^0.22.0"
SQLAlchemy = "^1.4.49"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
